Create InputReceiver.cpp
